pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "practical_ai_azure_keyvault"
version = "0.1.0"
description = "A package for managing Azure OpenAI configuration and authentication."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "azure-identity",
    "azure-keyvault-secrets",
    "openai"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.6"
